org-notmuch-{,search-,tree-}store-link: Simplify
No longer needlessly bind local variables.

No longer explicitly return the link string because (a) it currently
only matters that non-nil is returned and (b) if something more
meaningful is returned, then it should be the link *plist*, which is
what we do now.

The store-link abstraction sadly relies on side-effects, specifically
the variable `org-store-link-plist' is used to pass around values.
(There is a note in `org-store-link' in `ol.el', which suggests that
this might eventually be fixed to use return values instead.)

For the most part this can be ignored, but in the case of
`org-notmuch-store-link' we cannot do that.  That function uses
`org-link-email-description', which has to be called in a separate
step to set `:description' (because it relies on values already being
stored in the plist) and gets the plist from `org-store-link-plist'.

We then also have to ensure we return non-nil, and do so by returning
the value of `org-store-link-plist' for the reason given above.