org-notmuch-{search,tree}-open: Absorb their *-follow-link functions
Some other link types also have "open" and "follow-link" functions,
but in most of these cases there is actually a reason to do so.
Our "notmuch" type for example provides a Custom option that lets
the user swap in an alternative "follow-link" function.  But the
"notmuch-search" and "notmuch-tree" do not do any such thing.